rename images and icons
