+++ assets/icons/IconType.ts
@@ -1,0 +1,6 @@
+/* eslint-disable @typescript-eslint/no-require-imports */
+export const IconType = {
+  facebook: require('@/assets/icons/facebook.png'),
+  twitter: require('@/assets/icons/instagram.png'),
+  instagram: require('@/assets/icons/twiter.png')
+}
