+++ assets/images/ImageType.ts
@@ -1,0 +1,4 @@
+/* eslint-disable @typescript-eslint/no-require-imports */
+export const ImagesType = {
+  loginWallpaper: require('@/assets/images/login-wallpaper.jpg'),
+}
