+++ src/ui/pages/Concursantes/List/components/SocialMediaIcon.tsx
@@ -1,17 +1,11 @@
-/* eslint-disable @typescript-eslint/no-require-imports */
 import { Image, StyleSheet, TouchableOpacity, View } from 'react-native'
-
-const icons = {
-  facebook: require('@/assets/icons/facebook.png'),
-  twitter: require('@/assets/icons/instagram.png'),
-  instagram: require('@/assets/icons/twiter.png'),
-}
+import { IconType } from '@/assets/icons/IconType'
 
 export const SocialMediaIcons = ({ social }: { social: { facebook: string; twitter: string; instagram: string } }) => {
   const socialLinks = [
-    { name: 'facebook', url: social.facebook, icon: icons.facebook },
-    { name: 'twitter', url: social.twitter, icon: icons.twitter },
-    { name: 'instagram', url: social.instagram, icon: icons.instagram },
+    { name: 'facebook', url: social.facebook, icon: IconType.facebook },
+    { name: 'twitter', url: social.twitter, icon: IconType.twitter },
+    { name: 'instagram', url: social.instagram, icon: IconType.instagram },
   ]
 
   return (
